+++ src/ashley/machina_extensions/forum_permission/admin.py
@@ -1,0 +1,6 @@
+"""
+Import the admin classes from the django-machina's forum_permission application
+"""
+
+# pylint: disable=unused-wildcard-import,wildcard-import
+from machina.apps.forum_permission.admin import *  # noqa isort:skip